switched integration tests to use HTTPS.
this required generating a custom SSL certificate for our
dev server and installing it to apache/config folder in Redmine's Bitnami stack.

loaded SSL certificate from our dev server and saved it
to a file:

    openssl s_client -CApath /etc/ssl/certs/ -connect 173.247.200.254:443 < /dev/null | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' > ~/ta-dev-public.crt

then created a keystore using this file:

    keytool -import -alias 173.247.200.254 -keystore ta-dev-cacerts -file ta-dev-public.crt

I added this file to the repo. it is now loaded and provided to
 RedmineManager.