Add getter for project status
Note1: Redmine api does not allow reopen, close or archive projects.
So as, Project.setStatus has only effect on the project bean, but
it does not change the status in the redmine database.
An issue was been recorded about this (documented in the code).

Note 2: But we still need setStatus as RedmineJSONParser needs it.
A solution to avoid creating this setter could be use storage.set
instead result.setStatus, but it is not possible as storage.set
is not visible from RedmineJSONParser